Improve dashboard and insights maintainability
