+++ nextjs/src/components/dashboard-view.js
@@ -1,7 +1,7 @@
 'use client'
 
 import Link from 'next/link'
-import { useEffect, useState } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import { useRouter } from 'next/navigation'
 import { useAuth, useDataMode, useDataChanged } from '@/components/providers'
 import { ApiError, apiGet, apiPost } from '@/lib/apiClient'
@@ -14,7 +14,7 @@ import {
   demoInsightsSnapshot,
   demoSavingsGoals,
 } from '@/lib/demoData'
-import { getCategoryPresentation, getEntryVisual, getInitialsLabel } from '@/lib/financeVisuals'
+import { getEntryVisual, getInitialsLabel } from '@/lib/financeVisuals'
 import {
   buildActivityFeed,
   buildDailySpendDetailsFromExpenses,
@@ -27,11 +27,23 @@ import {
   isInMonth,
 } from '@/lib/financeUtils'
 import {
-  buildBudgetPressureHighlight as buildSharedBudgetPressureHighlight,
-  buildCategoryBudgetHealth,
+  buildDashboardLivePaths,
+  buildDemoRecentCashFlow,
+  buildDerivedCategoryCards,
+  buildSampleCategoryCards,
+  getBudgetCtaLabel,
+  getBudgetHintText,
+  getBudgetHudModel,
+  getCategoryCards,
+  getFirstName,
+  getPreviewCategories,
+  getTopSavingsGoal,
+  hasCategoryBudgets,
+  hasOverallMonthlyLimit,
+  mergeRowsById,
+} from '@/lib/dashboardModels'
+import {
   buildFinancialHealth,
-  buildOverallBudgetHealth as buildSharedOverallBudgetHealth,
-  getMonthProgressState as getSharedMonthProgressState,
 } from '@/lib/budgetHealth'
 import {
   getSavingsGoalAvatar,
@@ -49,250 +61,21 @@ import MonthPacingChart from '@/components/ui/MonthPacingChart'
 import TransactionDetailSheet from '@/components/ui/TransactionDetailSheet'
 
 const ACTIVITY_PREVIEW_LIMIT = 6
-const CATEGORY_PREVIEW_LIMIT = 5
+
+export {
+  buildDerivedCategoryCards,
+  getBudgetCtaLabel,
+  getBudgetHintText,
+  getBudgetHudModel,
+  getCategoryCards,
+  getMonthProgressState,
+  getBudgetPressureHighlight,
+} from '@/lib/dashboardModels'
 
 function getErrorMessage(error) {
   if (error instanceof ApiError) return error.message
   if (error instanceof Error && error.message) return error.message
   return 'Something went wrong while loading the live snapshot.'
-}
-
-function getFirstName(email) {
-  if (!email) return 'there'
-
-  const [name] = email.split('@')
-  const cleaned = name.replace(/[._-]+/g, ' ').trim()
-  return cleaned
-    .split(/\s+/)
-    .filter(Boolean)
-    .map((part) => part[0].toUpperCase() + part.slice(1))
-    .join(' ')
-}
-
-export function hasOverallMonthlyLimit(summary) {
-  return Number(summary?.monthly_limit ?? 0) > 0
-}
-
-function hasBudgetedCategoryStatuses(categoryStatuses) {
-  return Array.isArray(categoryStatuses)
-    && categoryStatuses.some((item) => Number(item?.monthly_limit ?? 0) > 0)
-}
-
-export function hasCategoryBudgets(summary) {
-  return hasBudgetedCategoryStatuses(summary?.category_statuses)
-}
-
-export function getBudgetCtaLabel(summary) {
-  if (hasOverallMonthlyLimit(summary)) return 'Edit budget'
-  if (hasCategoryBudgets(summary)) return 'Set overall limit'
-  return 'Set budget'
-}
-
-export function getBudgetHintText(summary) {
-  if (hasOverallMonthlyLimit(summary)) {
-    return `Current limit: ${formatCurrency(summary?.monthly_limit)}. Changes take effect immediately.`
-  }
-
-  if (hasCategoryBudgets(summary)) {
-    return 'Category budgets are already set. Add an overall monthly limit here to control the monthly cap and overall-budget alerts.'
-  }
-
-  return 'Set an overall monthly limit here to control the monthly cap and overall-budget alerts.'
-}
-
-function getSafeMoneyNumber(value) {
-  const amount = Number(value)
-  return Number.isFinite(amount) ? amount : 0
-}
-
-export function getMonthProgressState(month, { observedDayCount = 0, referenceDate = new Date() } = {}) {
-  return getSharedMonthProgressState(month, { observedDayCount, referenceDate })
-}
-
-export function getBudgetHudModel(summary, { month, observedDayCount = 0, referenceDate = new Date(), availability = summary ? 'ready' : 'loading' } = {}) {
-  const overallHealth = buildSharedOverallBudgetHealth({
-    summary,
-    availability,
-    month,
-    observedDayCount,
-    referenceDate,
-  })
-  const financialHealth = buildFinancialHealth({ summary, availability })
-  const budget = overallHealth.totalBudget ?? 0
-  const spent = overallHealth.spent ?? 0
-  const income = getSafeMoneyNumber(summary?.total_income)
-  const net = financialHealth.netAmount ?? (income != null && overallHealth.spent != null
-    ? Number((income - overallHealth.spent).toFixed(2))
-    : 0)
-  const hasBudget = overallHealth.key !== 'no_budget' && overallHealth.key !== 'loading' && overallHealth.key !== 'unavailable'
-
-  const metrics = overallHealth.key === 'loading' || overallHealth.key === 'unavailable'
-    ? [
-      { label: 'Spent', value: '--', hint: 'Current month' },
-      {
-        label: 'Days left',
-        value: overallHealth.monthState.daysRemaining ? String(overallHealth.monthState.daysRemaining) : '--',
-        hint: 'Including today',
-      },
-      { label: 'Daily allowance', value: '--', hint: 'Left per day' },
-      { label: 'Net this month', value: '--', hint: 'Income minus spend' },
-    ]
-    : [
-      {
-        label: 'Spent',
-        value: formatCurrency(spent),
-        hint: hasBudget ? `${Math.round(overallHealth.progressPercentage)}% of budget used` : 'Current month',
-      },
-      {
-        label: 'Days left',
-        value: overallHealth.daysRemaining == null ? '--' : String(overallHealth.daysRemaining),
-        hint: 'Including today',
-      },
-      {
-        label: 'Daily allowance',
-        value: overallHealth.dailyAllowance == null ? '--' : formatCurrency(overallHealth.dailyAllowance),
-        hint: overallHealth.key === 'over_budget' ? 'Needs correction' : 'Left per day',
-      },
-      {
-        label: 'Net this month',
-        value: financialHealth.netAmount == null ? '--' : formatCurrency(financialHealth.netAmount),
-        hint: financialHealth.key === 'negative_cash_flow'
-          ? 'Expenses above income'
-          : 'Income minus spend',
-      },
-    ]
-
-  return {
-    ...overallHealth,
-    badge: overallHealth.label,
-    value: overallHealth.primaryValue,
-    progressWidth: `${overallHealth.progressPercentage}%`,
-    budget,
-    spent,
-    income: income ?? 0,
-    remaining: overallHealth.remaining,
-    net,
-    daysRemaining: overallHealth.daysRemaining,
-    hasBudget,
-    isOverBudget: overallHealth.key === 'over_budget',
-    isNearLimit: overallHealth.key === 'near_limit',
-    metrics,
-  }
-}
-
-function buildLiveCategoryCards(categoryStatuses = []) {
-  return categoryStatuses.map((item) => {
-    const presentation = getCategoryPresentation({
-      name: item.category_name,
-      icon: item.category_icon,
-      kind: 'expense',
-    })
-    const amount = Number(item.spent ?? 0)
-    const categoryHealth = buildCategoryBudgetHealth({
-      monthlyLimit: item.monthly_limit,
-      spent: item.spent,
-      actualsAvailable: true,
-    })
-
-    return {
-      id: item.category_id ?? item.category_name ?? `${item.category_name}-${amount}`,
-      name: presentation.label,
-      symbol: presentation.symbol,
-      color: presentation.color,
-      soft: presentation.soft,
-      progress: categoryHealth.progressPercentage,
-      amount,
-      monthlyLimit: Number(item.monthly_limit ?? 0) > 0 ? Number(item.monthly_limit) : null,
-      remainingAmount: categoryHealth.remainingAmount,
-      note: categoryHealth.remainingText,
-      statusLabel: categoryHealth.label,
-      statusTone: categoryHealth.tone,
-    }
-  })
-}
-
-export function buildDerivedCategoryCards(expenses = []) {
-  const grouped = new Map()
-
-  expenses.forEach((expense) => {
-    const amount = Number(expense.amount ?? 0)
-    const key = expense.category_id ?? expense.category_name ?? 'uncategorized'
-    const displayName = expense.category_name
-      && String(expense.category_name).trim() !== ''
-      ? String(expense.category_name).trim()
-      : ''
-    const existing = grouped.get(key)
-    const current = existing ?? {
-      category_name: displayName,
-      category_icon: expense.category_icon ?? null,
-      total_amount: 0,
-      count: 0,
-    }
-    if (!current.category_icon && expense.category_icon) {
-      current.category_icon = expense.category_icon
-    }
-
-    current.total_amount += amount
-    current.count += 1
-    grouped.set(key, current)
-  })
-
-  const breakdown = Array.from(grouped.entries())
-    .map(([key, item]) => ({
-      category_id: key,
-      category_name: item.category_name,
-      category_icon: item.category_icon ?? null,
-      total_amount: Number(item.total_amount.toFixed(2)),
-      count: item.count,
-    }))
-    .sort((left, right) => right.total_amount - left.total_amount)
-
-  const totalExpenses = breakdown.reduce((sum, item) => sum + Number(item.total_amount ?? 0), 0)
-
-  return breakdown.map((item) => {
-    const presentation = getCategoryPresentation({
-      name: item.category_name,
-      icon: item.category_icon,
-      kind: 'expense',
-    })
-    const amount = Number(item.total_amount ?? 0)
-    const share = totalExpenses > 0 ? (amount / totalExpenses) * 100 : 0
-
-    return {
-      id: item.category_id ?? item.category_name ?? `${item.category_name}-${amount}`,
-      name: presentation.label,
-      symbol: presentation.symbol,
-      color: presentation.color,
-      soft: presentation.soft,
-      progress: Math.min(share, 100),
-      amount,
-      monthlyLimit: null,
-      remainingAmount: null,
-      note: `${Math.round(share) || 0}% of spend`,
-    }
-  })
-}
-
-export function getCategoryCards(categoryStatuses, expenses = [], derivedCategoryCards = null) {
-  return hasBudgetedCategoryStatuses(categoryStatuses)
-    ? buildLiveCategoryCards(categoryStatuses)
-    : Array.isArray(derivedCategoryCards) ? derivedCategoryCards : buildDerivedCategoryCards(expenses)
-}
-
-export function getBudgetPressureHighlight(summary, expenses = [], derivedCategoryCards = null) {
-  const spendShareCards = Array.isArray(derivedCategoryCards)
-    ? derivedCategoryCards
-    : buildDerivedCategoryCards(expenses)
-
-  return buildSharedBudgetPressureHighlight({
-    categoryStatuses: summary?.category_statuses,
-    fallbackSpendCards: spendShareCards.map((card) => ({
-      name: card.name,
-      note: card.note,
-      progress: card.progress,
-      amount: card.amount,
-    })),
-  })
 }
 
 function LiveNotice({ message, onRetry }) {
@@ -346,20 +129,29 @@ export default function DashboardView() {
         message: '',
       }))
 
+      const paths = buildDashboardLivePaths(currentMonth)
       const results = await Promise.allSettled([
-        apiGet(`/api/budget/summary?month=${encodeURIComponent(currentMonth)}`, {
+        apiGet(paths.summary, {
           accessToken: session.accessToken,
           signal: controller.signal,
         }),
-        apiGet('/api/expenses', {
+        apiGet(paths.cashFlowExpenses, {
           accessToken: session.accessToken,
           signal: controller.signal,
         }),
-        apiGet('/api/income', {
+        apiGet(paths.recentExpenses, {
           accessToken: session.accessToken,
           signal: controller.signal,
         }),
-        apiGet('/api/savings-goals', {
+        apiGet(paths.cashFlowIncome, {
+          accessToken: session.accessToken,
+          signal: controller.signal,
+        }),
+        apiGet(paths.recentIncome, {
+          accessToken: session.accessToken,
+          signal: controller.signal,
+        }),
+        apiGet(paths.savingsGoals, {
           accessToken: session.accessToken,
           signal: controller.signal,
         }),
@@ -378,9 +170,11 @@ export default function DashboardView() {
       }
 
       const summaryResult = results[0]
-      const expensesResult = results[1]
-      const incomeResult = results[2]
-      const savingsGoalsResult = results[3]
+      const cashFlowExpensesResult = results[1]
+      const recentExpensesResult = results[2]
+      const cashFlowIncomeResult = results[3]
+      const recentIncomeResult = results[4]
+      const savingsGoalsResult = results[5]
       const failedCount = results.filter((result) => result.status === 'rejected').length
 
       setLiveState({
@@ -391,8 +185,14 @@ export default function DashboardView() {
             : 'Some live sections are missing for the moment, but the rest of the month is still visible.'
           : '',
         summary: summaryResult.status === 'fulfilled' ? summaryResult.value : null,
-        expenses: expensesResult.status === 'fulfilled' ? expensesResult.value : [],
-        income: incomeResult.status === 'fulfilled' ? incomeResult.value : [],
+        expenses: mergeRowsById(
+          cashFlowExpensesResult.status === 'fulfilled' ? cashFlowExpensesResult.value : [],
+          recentExpensesResult.status === 'fulfilled' ? recentExpensesResult.value : []
+        ),
+        income: mergeRowsById(
+          cashFlowIncomeResult.status === 'fulfilled' ? cashFlowIncomeResult.value : [],
+          recentIncomeResult.status === 'fulfilled' ? recentIncomeResult.value : []
+        ),
         savingsGoals: savingsGoalsResult.status === 'fulfilled' ? savingsGoalsResult.value : null,
       })
     }
@@ -466,10 +266,6 @@ export default function DashboardView() {
     }
   }
 
-  if (!isReady || !session?.accessToken) {
-    return null
-  }
-
   const summary = isSampleMode ? demoBudgetSummary : liveState.summary
   const summaryAvailability = isSampleMode
     ? 'ready'
@@ -478,101 +274,79 @@ export default function DashboardView() {
       : liveState.status === 'loading'
         ? 'loading'
         : 'unavailable'
-  const currentMonthExpenses = isSampleMode
-    ? []
-    : liveState.expenses.filter((expense) => isInMonth(expense.date || expense.created_at, currentMonth))
-  const categoryTransactionDetails = isSampleMode
-    ? (demoInsightsSnapshot?.dailySpend?.details ?? [])
-    : buildDailySpendDetailsFromExpenses(currentMonthExpenses)
-  const activity = isSampleMode
-    ? demoActivity
-    : buildActivityFeed(liveState.expenses, liveState.income)
-  const filteredActivity = activityFilter === 'all'
-    ? activity
-    : activity.filter((entry) => entry.kind === activityFilter)
-  const recentActivity = filteredActivity.slice(0, ACTIVITY_PREVIEW_LIMIT)
+  const currentMonthExpenses = useMemo(() => (
+    isSampleMode
+      ? []
+      : liveState.expenses.filter((expense) => isInMonth(expense.date || expense.created_at, currentMonth))
+  ), [currentMonth, isSampleMode, liveState.expenses])
+  const categoryTransactionDetails = useMemo(() => (
+    isSampleMode
+      ? (demoInsightsSnapshot?.dailySpend?.details ?? [])
+      : buildDailySpendDetailsFromExpenses(currentMonthExpenses)
+  ), [currentMonthExpenses, isSampleMode])
+  const activity = useMemo(() => (
+    isSampleMode
+      ? demoActivity
+      : buildActivityFeed(liveState.expenses, liveState.income)
+  ), [isSampleMode, liveState.expenses, liveState.income])
+  const filteredActivity = useMemo(() => (
+    activityFilter === 'all'
+      ? activity
+      : activity.filter((entry) => entry.kind === activityFilter)
+  ), [activity, activityFilter])
+  const recentActivity = useMemo(() => (
+    filteredActivity.slice(0, ACTIVITY_PREVIEW_LIMIT)
+  ), [filteredActivity])
   const chartMonth = isSampleMode ? DEMO_MONTH : summary?.month || currentMonth
-  const hasBudgetedStatuses = hasBudgetedCategoryStatuses(summary?.category_statuses)
-  const derivedCategoryCards = !isSampleMode && !hasBudgetedStatuses
-    ? buildDerivedCategoryCards(currentMonthExpenses)
-    : null
-  const categoryCards = isSampleMode
-    ? demoCategoryBudgets.map((item) => {
-      const presentation = getCategoryPresentation({ name: item.name, kind: 'expense' })
-      const categoryHealth = buildCategoryBudgetHealth({
-        monthlyLimit: item.budget,
-        spent: item.spent,
-        actualsAvailable: true,
-      })
-      return {
-        id: item.id,
-        name: presentation.label,
-        symbol: presentation.symbol,
-        color: presentation.color,
-        soft: presentation.soft,
-        progress: categoryHealth.progressPercentage,
-        amount: item.spent,
-        monthlyLimit: item.budget,
-        remainingAmount: categoryHealth.remainingAmount,
-        note: categoryHealth.remainingText,
-        statusLabel: categoryHealth.label,
-        statusTone: categoryHealth.tone,
-      }
-    })
-    : getCategoryCards(summary?.category_statuses, currentMonthExpenses, derivedCategoryCards)
+  const hasBudgetedStatuses = hasCategoryBudgets(summary)
+  const derivedCategoryCards = useMemo(() => (
+    !isSampleMode && !hasBudgetedStatuses
+      ? buildDerivedCategoryCards(currentMonthExpenses)
+      : null
+  ), [currentMonthExpenses, hasBudgetedStatuses, isSampleMode])
+  const categoryCards = useMemo(() => (
+    isSampleMode
+      ? buildSampleCategoryCards(demoCategoryBudgets)
+      : getCategoryCards(summary?.category_statuses, currentMonthExpenses, derivedCategoryCards)
+  ), [currentMonthExpenses, derivedCategoryCards, isSampleMode, summary?.category_statuses])
   const budgetCtaLabel = getBudgetCtaLabel(summary)
-  const trendPoints = isSampleMode
-    ? demoBudgetTrend
-    : buildMonthlySpendTrend(liveState.expenses, currentMonth)
-  const hudState = getBudgetHudModel(summary, {
+  const trendPoints = useMemo(() => (
+    isSampleMode
+      ? demoBudgetTrend
+      : buildMonthlySpendTrend(liveState.expenses, currentMonth)
+  ), [currentMonth, isSampleMode, liveState.expenses])
+  const hudState = useMemo(() => getBudgetHudModel(summary, {
     month: chartMonth,
     observedDayCount: trendPoints.length,
     availability: summaryAvailability,
-  })
-  const financialHealth = buildFinancialHealth({
+  }), [chartMonth, summary, summaryAvailability, trendPoints.length])
+  const financialHealth = useMemo(() => buildFinancialHealth({
     summary,
     availability: summaryAvailability,
-  })
+  }), [summary, summaryAvailability])
   const savingsGoals = isSampleMode ? demoSavingsGoals : liveState.savingsGoals
-  const topSavingsGoal = [...(savingsGoals?.goals ?? [])]
-    .filter((goal) => !goal.archived)
-    .sort((left, right) => {
-      const leftComplete = left.budget_context?.status === 'complete'
-      const rightComplete = right.budget_context?.status === 'complete'
-      if (leftComplete !== rightComplete) return leftComplete ? 1 : -1
-      const severity = { over_budget: 5, overdue: 5, tight: 4, no_budget: 3, ready: 2, complete: 1 }
-      const severityDifference = (severity[right.budget_context?.status] ?? 0) - (severity[left.budget_context?.status] ?? 0)
-      if (severityDifference) return severityDifference
-      return String(left.target_date).localeCompare(String(right.target_date))
-    })[0] ?? null
-  const recentCashFlow = isSampleMode
-    ? (() => {
-      const monthlyIncome = getSafeMoneyNumber(demoBudgetSummary?.total_income)
-      const monthlyExpenses = getSafeMoneyNumber(demoBudgetSummary?.total_expenses)
-      return [
-        { month: '2026-01-01', label: 'Jan', incomeAmount: monthlyIncome * 0.95, expenseAmount: monthlyExpenses * 0.88, netAmount: Number((monthlyIncome * 0.95 - monthlyExpenses * 0.88).toFixed(2)) },
-        { month: '2026-02-01', label: 'Feb', incomeAmount: monthlyIncome * 1.02, expenseAmount: monthlyExpenses * 1.05, netAmount: Number((monthlyIncome * 1.02 - monthlyExpenses * 1.05).toFixed(2)) },
-        { month: '2026-03-01', label: 'Mar', incomeAmount: monthlyIncome, expenseAmount: monthlyExpenses, netAmount: Number((monthlyIncome - monthlyExpenses).toFixed(2)) },
-      ]
-    })()
-    : buildRecentCashFlow(liveState.expenses, liveState.income, chartMonth, 3)
+  const topSavingsGoal = useMemo(() => (
+    getTopSavingsGoal(savingsGoals)
+  ), [savingsGoals])
+  const recentCashFlow = useMemo(() => (
+    isSampleMode
+      ? buildDemoRecentCashFlow(demoBudgetSummary)
+      : buildRecentCashFlow(liveState.expenses, liveState.income, chartMonth, 3)
+  ), [chartMonth, isSampleMode, liveState.expenses, liveState.income])
   const firstName = getFirstName(session?.user?.email)
   const periodLabel = formatMonthPeriod(chartMonth)
-  const previewCategories = [...categoryCards]
-    .sort((left, right) => {
-      const leftBudgeted = left.monthlyLimit != null && Number(left.monthlyLimit) > 0
-      const rightBudgeted = right.monthlyLimit != null && Number(right.monthlyLimit) > 0
-      if (leftBudgeted && rightBudgeted) {
-        return (Number(right.progress) || 0) - (Number(left.progress) || 0)
-      }
-      if (leftBudgeted !== rightBudgeted) return leftBudgeted ? -1 : 1
-      return (Number(right.amount) || 0) - (Number(left.amount) || 0)
-    })
-    .slice(0, CATEGORY_PREVIEW_LIMIT)
+  const previewCategories = useMemo(() => (
+    getPreviewCategories(categoryCards)
+  ), [categoryCards])
   const monthMarkerLabel = hudState.monthState?.monthLength
     ? `Today · Day ${hudState.monthState.activeDay}`
     : null
   const chartSpendValue = trendPoints.length ? formatCurrency(trendPoints.at(-1) ?? 0) : '--'
+
+  if (!isReady || !session?.accessToken) {
+    return null
+  }
+
   return (
     <>
     <section className="app-screen dashboard-screen">
